+++ src/backend/api/Fusion.Resources.Api/Controllers/Validation/CustomValidatorExtensions.cs
@@ -190,42 +190,6 @@ namespace Fusion.Resources.Api.Controllers
         }
     }
 
-    public class EmailDomainValidator : PropertyValidator, IPropertyValidator
-    {
-        private static string[] InvalidDomains = new[]
-        {
-            "gmail.com",
-            "hotmail.com",
-            "aol.com",
-            "ymail.com",
-            "yahoo.no",
-            "yahoo.com",
-            "outlook.com",
-            "zohomail.eu",
-            "mail.com",
-            "protonmail.com",
-            "icloud.com",
-            "gmx.com",
-            "gmx.us",
-            "yandex.com"
-        };
-
-        protected override bool IsValid(PropertyValidatorContext context)
-        {
-            var value = context.PropertyValue as string;
-
-            if (!string.IsNullOrEmpty(value) && value.Contains("@")) 
-            {
-                var domain = value.Split("@").Last();
-                if (InvalidDomains.Contains(domain, StringComparer.OrdinalIgnoreCase))
-                    return false;
-            }
-
-            return true;
-        }
-        
-    }
-
     
 
 }
